7: add POST comment to article

+++ models/article-model.js
@@ -18,7 +18,7 @@ exports.fetchArticleByID = (articleID) => {
 
 exports.fetchAllArticles = () =>{
     return db
-    .query(`SELECT articles.article_id, articles.title, articles.topic, articles.author, articles.created_at, articles.votes, articles.article_img_url, COUNT(comments.article_id) AS comment_count FROM articles LEFT JOIN comments ON comments.article_id =articles.article_id
+    .query(`SELECT articles.article_id, articles.title, articles.topic, articles.author, articles.created_at, articles.votes, articles.article_img_url, COUNT(comments.article_id)::INT AS comment_count FROM articles LEFT JOIN comments ON comments.article_id =articles.article_id
         GROUP BY 
         articles.article_id, 
         articles.title,
